Update to v2014-0205. Yes, I'm incredibly late.
A quick note regarding In-Home Streaming support. There's only one style
for the drop down indicator/arrow for streaming so there's no way for me
to make it blue when next to the install button and green when next to
the play button. I guess that's due to the default interface not having
a different colour for those buttons, but either way I don't find it
looks bad. It's visible, gets the job done, and I don't really see any
other way around it.

Changes in this version
• Preliminary In-Home Streaming beta support.
• Added locale.ini support for Ukrainian users.
• Game Title moved to the right of the Overlay in the 'Low Resolution
Overlay' files.
• Added +3 & +4 font size tweaks.