Preliminary Fix For Music Beta
I'm not going to have any more time to spend on the update today, so
I've put together an archive with the current build. The major bugs are
gone at least, so it should be pretty usable. The music player pop-up
isn't finished yet, but that's about it.
--Pulseh
(http://steamcommunity.com/groups/pixelvisionskin#announcements/detail/268280328684475058)