Update to v2014-0925
I skipped a bunch of updates, as follows. So far as I can tell, the
current update is for non-beta. I'll check it out and push again if it
breaks in the beta client.

Skipped updates:
v2014-0726 CLIENT BETA   (26th of July 2014)
v2014-0808 CLIENT BETA   (8th of August 2014)
v2014-0904 CLIENT BETA   (4th of September 2014)